Mejoras generales en #40 #39 #32 #30 #29 #28 #27

+++ rolsac2-persistence/src/main/java/es/caib/rolsac2/persistence/model/JListaDocumentos.java
@@ -12,13 +12,13 @@ public class JListaDocumentos {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "lista-doc-sequence")
     @Column(name = "LSDO_CODIGO", nullable = false)
-    private Integer codigo;
+    private Long codigo;
 
-    public Integer getCodigo() {
+    public Long getCodigo() {
         return codigo;
     }
 
-    public void setCodigo(Integer id) {
+    public void setCodigo(Long id) {
         this.codigo = id;
     }
